requirements.txt
fal-client
supabase
python-dotenv
pillow
httpx
aiofiles
requests
beautifulsoup4
